fix phonons.py into phonons/plotly.py and phonons/helpers.py
