+++ pymatviz/phonons/__init__.py
@@ -1,0 +1,6 @@
+"""Plotting functions for phonon band structures and density of states. Supports
+pymatgen PhononDos and PhononBandStructureSymmLine as well as phonopy TotalDos and
+BandStructure objects.
+"""
+
+from pymatviz.phonons.plotly import phonon_bands, phonon_bands_and_dos, phonon_dos
